more reducing warnings

+++ freex3d/src/lib/vrml_parser/CParseParser.c
@@ -239,7 +239,7 @@ BOOL parseType(struct VRMLParser* me, int type,   union anyVrml *defaultVal) {
 
 
 /* put the string value of the PROTO field into the input stream */
-void replaceProtoField(struct VRMLLexer *me, struct ProtoDefinition *thisProto, char *thisID, char **outTextPtr, int *outSize) {
+void replaceProtoField(struct VRMLLexer *me, struct ProtoDefinition *thisProto, char *thisID, char **outTextPtr, size_t *outSize) {
     struct ProtoFieldDecl* pdecl=NULL;
     /* find the ascii name, and try and find it's protodefinition by type */
 #ifdef CPARSERVERBOSE
